Resources.zh-TW.resx
I will make zh-CN soon（Simplified Chinese）,it have some different word because regional difference. For example , bus in Taiwan called '公車‘ ,but it called '巴士' in Hong Kong and it called '公交車' in mainland china. But I think the difference of taiwan and Hongkong is small so i would make zh-TW and zh-CN only
If it have some problem please connect me on twitter @sp1900yellow